fix float locale issue in ncnn2mem, fix long line parsing in ncnn2int8
